### Context

The cleanup function in testing-library/react should be called automaticly. However it is never called, resulting in performance issues (memory leaks ?) and sometimes invalid tests.

### Solution

testing-library registers cleanup via afterEach, which is not available in test-setup, only on test-setup-after-env
We import testing-library in test-setup via storybook mocks, as we add decorators/parameters

The solution is to move the import in the `storiesOf` which should be called only once on each story test file. This way, in the test-setup phase, testing-library is no longer imported, and afterEach will be available when `storiesOf` is called
This also has another benefits: for tests not using testing-library, it will never be imported.

Perf test using hyperfine:

Impact on time on react (components / warm cache): (no change)
- before: 36.123 s ±  0.649 
- after: 35.024 s ±  1.494 s    


Also do that on react-native. We don't seem to have a problem, but there is a cleanup function too.
Impact on time on react-native (instructors-app / warm cache): (no change)
- before: 205s
- after: 202s